图表界面修改，chart文件夹换为chart1文件夹
