+++ lib/homepage.dart
@@ -4,12 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_jizhangapp/login/password_reset.dart';
 import './login/graphical_password_register.dart';
 import './login/password_confirm.dart';
-import './login/register.dart';
-import 'chart/chartpage.dart';
-import 'chart/select.dart';
-import 'login/login.dart';
-import 'service/shared_pref.dart';
-import 'unknown_page.dart';
+import 'chart1/chartpage.dart';
 import 'editbill/edit_bill_page.dart';
 import 'login/remove_user_data.dart';
 import './total/TotalPage.dart';
